feat: move faker to devdependencies
Move faker.js to devDependencies as it's not used in production app. Also add node args to pm2
config to support es6
