+++ package.json
@@ -11,7 +11,7 @@
     "node": ">=14.0.0"
   },
   "scripts": {
-    "start": "pm2 start ecosystem.config.json --no-daemon",
+    "start": "tsc --build && eslint . --fix && pm2 start ecosystem.config.json --no-daemon",
     "compile": "tsc --build && eslint . --fix",
     "compile:watch": "tsc --build --watch",
     "pre:dev": "cross-env NODE_ENV=development nodemon --experimental-modules --es-module-specifier-resolution=node dist/index.js",
@@ -58,6 +58,7 @@
   "devDependencies": {
     "@commitlint/cli": "^16.2.3",
     "@commitlint/config-conventional": "^16.2.1",
+    "@faker-js/faker": "^6.2.0",
     "@jest/globals": "^27.5.1",
     "@types/bcryptjs": "^2.4.2",
     "@types/compression": "^1.7.2",
@@ -100,7 +101,6 @@
     "typescript": "^4.5.4"
   },
   "dependencies": {
-    "@faker-js/faker": "^6.0.0-alpha.7",
     "bcryptjs": "^2.4.3",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
