+++ src/modules/token/index.ts
@@ -1,7 +1,6 @@
-import * as tokenFixture from './token.fixture';
 import * as tokenService from './token.service';
 import Token from './token.model';
 import * as tokenInterfaces from './token.interfaces';
 import tokenTypes from './token.types';
 
-export { tokenFixture, tokenService, Token, tokenInterfaces, tokenTypes };
+export { tokenService, Token, tokenInterfaces, tokenTypes };
