+++ src/modules/token/token.model.test.ts
@@ -1,10 +1,25 @@
 import moment from 'moment';
+import mongoose from 'mongoose';
+import { faker } from '@faker-js/faker';
 import config from '../../config/config';
 import { NewToken } from './token.interfaces';
-import { userOne } from '../user/user.fixture';
-import { userOneAccessToken } from './token.fixture';
 import tokenTypes from './token.types';
 import Token from './token.model';
+import * as tokenService from './token.service';
+
+const password = 'password1';
+const accessTokenExpires = moment().add(config.jwt.accessExpirationMinutes, 'minutes');
+
+const userOne = {
+  _id: new mongoose.Types.ObjectId(),
+  name: faker.name.findName(),
+  email: faker.internet.email().toLowerCase(),
+  password,
+  role: 'user',
+  isEmailVerified: false,
+};
+
+const userOneAccessToken = tokenService.generateToken(userOne._id, accessTokenExpires, tokenTypes.ACCESS);
 
 describe('Token Model', () => {
   const refreshTokenExpires = moment().add(config.jwt.refreshExpirationDays, 'days');
