+++ src/modules/user/index.ts
@@ -1,8 +1,7 @@
 import * as userController from './user.controller';
-import * as userFixture from './user.fixture';
 import * as userInterfaces from './user.interfaces';
 import User from './user.model';
 import * as userService from './user.service';
 import * as userValidation from './user.validation';
 
-export { userController, userFixture, userInterfaces, User, userService, userValidation };
+export { userController, userInterfaces, User, userService, userValidation };
