--- src/modules/user/user.fixture.ts
@@ -1,39 +1,0 @@
-import mongoose from 'mongoose';
-import bcrypt from 'bcryptjs';
-import { faker } from '@faker-js/faker';
-import User from './user.model';
-
-const password = 'password1';
-const salt = bcrypt.genSaltSync(8);
-const hashedPassword = bcrypt.hashSync(password, salt);
-
-export const userOne = {
-  _id: new mongoose.Types.ObjectId(),
-  name: faker.name.findName(),
-  email: faker.internet.email().toLowerCase(),
-  password,
-  role: 'user',
-  isEmailVerified: false,
-};
-
-export const userTwo = {
-  _id: new mongoose.Types.ObjectId(),
-  name: faker.name.findName(),
-  email: faker.internet.email().toLowerCase(),
-  password,
-  role: 'user',
-  isEmailVerified: false,
-};
-
-export const admin = {
-  _id: new mongoose.Types.ObjectId(),
-  name: faker.name.findName(),
-  email: faker.internet.email().toLowerCase(),
-  password,
-  role: 'admin',
-  isEmailVerified: false,
-};
-
-export const insertUsers = async (users: Record<string, any>[]) => {
-  await User.insertMany(users.map((user) => ({ ...user, password: hashedPassword })));
-};
